// peniel-crm-front/src/app/models/contact.model.ts
export class Contact {

  raisonSociale: String;
  siren: String;
  activitePrincipale: String;
  typePrestation: String;
  typeContact: String;
  id?: String;
  nom: string;
  prenom;
  telephoneMobile: String;
  siteInternet: String;
  creationDate?: Date;
  assigne: string;
  telephoneFixe: string;
  telephoneMobile1: string;
  telephoneMobile2: string;
  emailPerso: string;
  emailPro: string;
  structure: string;
  activite: string;
  adressePerso: string;
  adressePro: string;
  adressePersoCodePostal: string;
  adresseProCodePostal: string;
  adressePersoVille: string;
  adresseProVille: string;
  adressePersoPays: string;
  adresseProPays: string;
  adressePersoZoneGeo: string;
  adresseProZoneGeo: string;
  adressePersoZoneGeoSuite: string;
  adresseProZoneGeoSuite: string;
  debutAbonnement: string;
  renAbonnement: string;
  source: string;
  typeAbonnement: string;
  notes: string;
  naissance: string;
  nationalite: string;
  civilite: string;
  statut: string;
  fonction: string;
  situationFamiliale: string;
  abonne: string;
  typeHebergementPrefere: string;
  typeRestaurantPrefere: string;
  tailleEntreprise: string;
  contactPrincipalNom: string;
  contactPrincipalTelephoneMobile: string;
  contactSecondaireTelephoneMobile: string;
  contactPrincipalEmail: string;
  contactSecondaireEmail: string;
  beneficiaires: []
}